+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNXVFNVN1ABN977B3K030E2W"
+        "01KP0KM9ASEY9VW0351DHNEBAX"
     )
     r"""The unique identifier of the data source"""